Declared the variables
Initilaized a list with some words so that system can Randomly pick up a word for the game.

LETTER_NOT_FOUND_CHARACTER is used to place an underscore for the missing letteers

EMPTY_CHAR_CHECK holds a char that returns a null value. This will be used as check in order to place the LETTER_NOT_FOUND_CHAR

3 different loops used in this program
2 while loops
one to keep the game on going until the player decides to stop
another to keep loopingn as long as the players hasn t guessed the right word and that the player's has still enough chances.

for loop to check the random letter index by index in order to place the right char in the right index in the _userFoundChars array

also a foreach loop to loop through the _userFoundChars and displaying letters found or an _ for empty indexes.